Add files via upload
The "Basic Info" tab is the one to be filled out in the FinalProject excel sheet. The rest is going to be populated automatically. The support text in the "ListOfFunctions" is not in the final state yet. This is not to say that it is bad. I removed the password protecting the worksheets to make it easier.